Revert "add new tokenizer system support (#1575)"
This reverts commit 1a44a8bcc7c48b178234625faeff09d14de4f67f.
