+++ src/megatron/bridge/data/builders/finetuning_dataset.py
@@ -18,7 +18,6 @@ from typing import Any, Optional, Union
 
 import torch
 from megatron.core.msc_utils import MultiStorageClientFeature
-from megatron.core.tokenizers.text.libraries import HuggingFaceTokenizer
 
 from megatron.bridge.data.datasets.packed_sequence import PackedSequenceSpecs
 from megatron.bridge.data.datasets.sft import create_sft_dataset
@@ -314,23 +313,10 @@ class FinetuningDatasetBuilder:
 
     def _extract_tokenizer_model_name(self) -> str:
         """Automatically get the model name from model path."""
-        # Legacy tokenizer compatibility
-        if getattr(self.tokenizer, "legacy", False):
-            tokenizer_cls = _HuggingFaceTokenizer
-            tokenizer_instance = self.tokenizer
-        else:
-            tokenizer_cls = HuggingFaceTokenizer
-            tokenizer_instance = self.tokenizer._tokenizer
-
         if self.packed_sequence_specs and self.packed_sequence_specs.tokenizer_model_name is not None:
             return self.packed_sequence_specs.tokenizer_model_name
-        elif isinstance(tokenizer_instance, tokenizer_cls):
-            # Legacy tokenizer compatibility
-            if getattr(self.tokenizer, "legacy", False):
-                name = self.tokenizer._tokenizer.name_or_path
-            else:
-                name = self.tokenizer.path
-
+        elif isinstance(self.tokenizer, _HuggingFaceTokenizer):
+            name = self.tokenizer._tokenizer.name_or_path
             if name.endswith("context/nemo_tokenizer"):
                 # NEMO_HOME/hf_org/hf_model/context/nemo_tokenizer => hf_org--hf_model
                 tokenizer_model_name = "--".join(name.split("/")[-4:-2])
