+++ src/megatron/bridge/data/datasets/sft.py
@@ -33,7 +33,6 @@ from megatron.bridge.data.datasets.utils import (
     _JSONLMemMapDataset,
     _OnlineSampleMapping,
     _preprocess,
-    _tokenize,
 )
 from megatron.bridge.training.tokenizers.tokenizer import MegatronTokenizer
 
@@ -589,7 +588,7 @@ class GPTSFTDataset(Dataset):
                     raise e
 
         template_strings, template_strings_keys = self._separate_template(prompt_template_values)
-        template_ids = [_tokenize(self.tokenizer, s) for s in template_strings]
+        template_ids = [self.tokenizer.text_to_ids(s) for s in template_strings]
         context_ids, answer_ids = self._multiple_truncation(template_ids, template_strings_keys)
 
         if self.virtual_tokens:
@@ -1075,16 +1074,16 @@ class GPTSFTChatDataset(GPTSFTDataset):
             LABEL_START = self.special_tokens["label_start"]
             END_NAME_SIGNAL = self.special_tokens["end_of_name"]
 
-            id1 = _tokenize(self.tokenizer, PREFIX_STR)
-            id2 = _tokenize(self.tokenizer, PREFIX_STR + LABEL_START)
+            id1 = self.tokenizer.text_to_ids(PREFIX_STR)
+            id2 = self.tokenizer.text_to_ids(PREFIX_STR + LABEL_START)
             self.label_start_tokens = id2[len(id1) :]
 
-            id1 = _tokenize(self.tokenizer, PREFIX_STR + END_NAME_SIGNAL)
-            id2 = _tokenize(self.tokenizer, PREFIX_STR)
+            id1 = self.tokenizer.text_to_ids(PREFIX_STR + END_NAME_SIGNAL)
+            id2 = self.tokenizer.text_to_ids(PREFIX_STR)
             self.name_end_token_ids = id1[len(id2) :]
 
-            id1 = _tokenize(self.tokenizer, PREFIX_STR + self.special_tokens["turn_start"])
-            id2 = _tokenize(self.tokenizer, PREFIX_STR)
+            id1 = self.tokenizer.text_to_ids(PREFIX_STR + self.special_tokens["turn_start"])
+            id2 = self.tokenizer.text_to_ids(PREFIX_STR)
             self.num_turn_start_tokens = len(id1) - len(id2)
 
     def _process_example(self, example):
