+++ src/megatron/bridge/training/tokenizers/config.py
@@ -21,12 +21,6 @@ from typing import Any, Literal, Optional, Union
 class TokenizerConfig:
     """Configuration settings for the tokenizer."""
 
-    legacy_tokenizer: Optional[bool] = False
-    """To use Megatron-Bridge legacy tokenizer system."""
-
-    metadata_path: Optional[Union[str | dict]] = None
-    """Path to the tokenizer metadata file."""
-
     vocab_size: Optional[int] = None
     """Size of vocab before EOD or padding."""
 
@@ -58,12 +52,6 @@ class TokenizerConfig:
     tokenizer_model: Optional[Union[str, Path]] = None
     """Sentencepiece tokenizer model or the `pretrained_model_name_or_path` for a HuggingFace tokenizer."""
 
-    special_tokens: Optional[list[str]] = None
-    """List of special tokens. For TikToken, needs to have ["<unk>", "<s>", "</s>"]"""
-
-    chat_template: Optional[str] = None
-    """Custom chat template in jinja format for conversation formatting"""
-
     tiktoken_pattern: Optional[str] = None
     """Which tiktoken pattern to use. Options: [v1, v2]"""
 
@@ -74,6 +62,7 @@ class TokenizerConfig:
     """List of tiktoken special tokens, needs to have ["<unk>", "<s>", "</s>"]"""
 
     tokenizer_prompt_format: Optional[str] = None
+    special_tokens: Optional[list[str]] = None
     image_tag_type: Optional[str] = None
 
     hf_tokenizer_kwargs: dict[str, Any] | None = field(default_factory=dict)
@@ -82,26 +71,12 @@ class TokenizerConfig:
     Common options include:
         - use_fast (bool): Whether to use fast tokenizer implementation
         - trust_remote_code (bool): Whether to trust remote code when loading tokenizer
-        - include_special_tokens (bool): Whether to include special tokens when converting text to ids
+        - chat_template (str): Custom chat template string for conversation formatting
 
     Example:
         hf_tokenizer_kwargs = {
             "use_fast": True,
             "trust_remote_code": True,
-            "include_special_tokens": True
-        }
-    """
-
-    sp_tokenizer_kwargs: dict[str, Any] | None = field(default_factory=dict)
-    """Additional keyword arguments to pass to SentencePiece tokenizer.
-
-    Common options include:
-        - legacy (bool): Whether to use legacy format of sentencepiece tokenizer
-        - ignore_extra_whitespaces (bool): Whether to ignore extra whitespaces in the input text while encoding
-
-    Example:
-        sp_tokenizer_kwargs = {
-            "legacy": True,
-            "ignore_extra_whitespaces": False,
+            "chat_template": "custom_template_string"
         }
     """
