+++ tests/functional_tests/training/test_inprocess_restart.py
@@ -135,7 +135,7 @@ def build_test_config(
         ),
         tokenizer=TokenizerConfig(
             tokenizer_type="NullTokenizer",
-            vocab_size=131072,
+            vocab_size=10000,
         ),
         checkpoint=CheckpointConfig(
             save=save_dir,
